+++ components/core/src/main/java/org/mallfoundry/district/rest/CityResponse.java
@@ -42,7 +42,7 @@ public class CityResponse {
         return this.city.getName();
     }
 
-    public Integer getPosition() {
+    public long getPosition() {
         return this.city.getPosition();
     }
 }
