+++ components/core/src/main/java/org/mallfoundry/district/rest/RegionResponse.java
@@ -42,7 +42,7 @@ public class RegionResponse {
         return this.region.getName();
     }
 
-    public Integer getPosition() {
+    public long getPosition() {
         return this.region.getPosition();
     }
 }
